# Handle Vsap thermal paper in central scan properly