--- a/apps/central-scan/backend/src/store.ts
+++ b/apps/central-scan/backend/src/store.ts
@@ -5,7 +5,7 @@
 import { Client as DbClient } from '@votingworks/db';
 import {
   AdjudicationStatus,
-  BallotPaperSize,
+  HmpbBallotPaperSize,
   BallotSheetInfo,
   BatchInfo,
   Iso8601Timestamp,
@@ -374,11 +374,11 @@ export class Store {
     );
   }
 
-  getBallotPaperSizeForElection(): BallotPaperSize {
+  getBallotPaperSizeForElection(): HmpbBallotPaperSize {
     const electionRecord = this.getElectionRecord();
     return (
       electionRecord?.electionDefinition.election.ballotLayout.paperSize ??
-      BallotPaperSize.Letter
+      HmpbBallotPaperSize.Letter
     );
   }