--- a/apps/design/frontend/src/ballot_screen.tsx
+++ b/apps/design/frontend/src/ballot_screen.tsx
@@ -3,7 +3,7 @@ import { assertDefined, range } from '@votingworks/basics';
 import {
   getPrecinctById,
   getBallotStyle,
-  BallotPaperSize,
+  HmpbBallotPaperSize,
   BallotType,
   getPartyForBallotStyle,
   BallotStyleId,
@@ -195,13 +195,13 @@ function PdfViewer({ pdfData }: { pdfData?: Buffer }) {
   );
 }
 
-export const paperSizeLabels: Record<BallotPaperSize, string> = {
-  [BallotPaperSize.Letter]: '8.5 x 11 inches (Letter)',
-  [BallotPaperSize.Legal]: '8.5 x 14 inches (Legal)',
-  [BallotPaperSize.Custom17]: '8.5 x 17 inches',
-  [BallotPaperSize.Custom18]: '8.5 x 18 inches',
-  [BallotPaperSize.Custom21]: '8.5 x 21 inches',
-  [BallotPaperSize.Custom22]: '8.5 x 22 inches',
+export const paperSizeLabels: Record<HmpbBallotPaperSize, string> = {
+  [HmpbBallotPaperSize.Letter]: '8.5 x 11 inches (Letter)',
+  [HmpbBallotPaperSize.Legal]: '8.5 x 14 inches (Legal)',
+  [HmpbBallotPaperSize.Custom17]: '8.5 x 17 inches',
+  [HmpbBallotPaperSize.Custom18]: '8.5 x 18 inches',
+  [HmpbBallotPaperSize.Custom21]: '8.5 x 21 inches',
+  [HmpbBallotPaperSize.Custom22]: '8.5 x 22 inches',
 };
 
 export function BallotScreen(): JSX.Element | null {