--- a/libs/ballot-interpreter/src/bmd/utils/qrcode.ts
+++ b/libs/ballot-interpreter/src/bmd/utils/qrcode.ts
@@ -9,9 +9,6 @@ import { Optional, Result, err, ok, assertDefined } from '@votingworks/basics';
 import { DetectedQrCode } from '../types';
 import { stats, Stats } from './luminosity';
 
-const LETTER_WIDTH_TO_HEIGHT_RATIO = 8.5 / 11;
-const LEGAL_WIDTH_TO_HEIGHT_RATIO = 8.5 / 14;
-
 const debug = makeDebug('ballot-interpreter:bmd:qrcode');
 
 /**
@@ -30,119 +27,14 @@ function decodeBase64FromUtf8(utf8StringData: Buffer): Buffer {
 export function* getSearchAreas(
   size: Size
 ): Generator<{ position: 'top' | 'bottom'; bounds: Rect }> {
-  const widthToHeightRatio = size.width / size.height;
-  const isLetter =
-    Math.abs(widthToHeightRatio - LETTER_WIDTH_TO_HEIGHT_RATIO) <
-    Math.abs(widthToHeightRatio - LEGAL_WIDTH_TO_HEIGHT_RATIO);
-
-  // If we're not letter size then the size of the scanned image may be too
-  // large compared to the actual paper size. Instead of looking at the top
-  // right and bottom left we look at the top right and bottom left, but with a
-  // smaller area.
-  if (!isLetter) {
-    const expectedLetterHeight = Math.round(
-      size.width / LETTER_WIDTH_TO_HEIGHT_RATIO
-    );
-    // QR code for BMD is top right, so appears top right or bottom left
-    const bmdWidth = Math.round((size.width * 2) / 5);
-    const bmdHeight = Math.round((expectedLetterHeight * 2) / 5);
-    // We look at the bottom first because we're assuming people will mostly
-    // scan sheets so they appear right-side up to them, but bottom-side first
-    // to the scanner.
-
-    // ┌───┐
-    // │   │
-    // ├─┐ │
-    // │ │ │
-    // └─┴─┘
-    yield {
-      position: 'bottom',
-      bounds: {
-        x: 0,
-        y: size.height - bmdHeight,
-        width: bmdWidth,
-        height: bmdHeight,
-      },
-    };
-
-    // ┌───┐
-    // ├─┐ │
-    // │ │ │
-    // ├─┘ │
-    // └───┘
-    yield {
-      position: 'bottom',
-      bounds: {
-        x: 0,
-        y: expectedLetterHeight - bmdHeight,
-        width: bmdWidth,
-        height: bmdHeight,
-      },
-    };
-
-    // ┌─┬─┐
-    // │ │ │
-    // │ └─┤
-    // │   │
-    // └───┘
-    yield {
-      position: 'top',
-      bounds: {
-        x: size.width - bmdWidth,
-        y: 0,
-        width: bmdWidth,
-        height: bmdHeight,
-      },
-    };
-
-    // ┌───┐
-    // │ ┌─┤
-    // │ │ │
-    // │ └─┤
-    // └───┘
-    yield {
-      position: 'top',
-      bounds: {
-        x: size.width - bmdWidth,
-        y: size.height - expectedLetterHeight,
-        width: bmdWidth,
-        height: bmdHeight,
-      },
-    };
-  }
-
-  // QR code for BMD is top right, so appears top right or bottom left
-  const bmdWidth = Math.round((size.width * 2) / 5);
-  const bmdHeight = Math.round((size.height * 2) / 5);
-  // We look at the bottom first because we're assuming people will mostly scan
-  // sheets so they appear right-side up to them, but bottom-side first to the
-  // scanner.
-
-  // ┌───┐
-  // │   │
-  // ├─┐ │
-  // └─┴─┘
-  yield {
-    position: 'bottom',
-    bounds: {
-      x: 0,
-      y: size.height - bmdHeight,
-      width: bmdWidth,
-      height: bmdHeight,
-    },
-  };
-
-  // ┌─┬─┐
-  // │ └─┤
-  // │   │
-  // └───┘
+  // TODO (#4980) Be more selective about bmd QR code search areas after merging with the hmpb and ensuring images from central scan are properly cropped.
   yield {
     position: 'top',
     bounds: {
-      x: size.width - bmdWidth,
+      x: 0,
       y: 0,
-      width: bmdWidth,
-      height: bmdHeight,
+      width: size.width,
+      height: size.height,
     },
   };
 }