--- a/libs/hmpb/src/preview/browser_preview.ts
+++ b/libs/hmpb/src/preview/browser_preview.ts
@@ -1,6 +1,6 @@
 import './polyfills';
 import {
-  BallotPaperSize,
+  HmpbBallotPaperSize,
   BallotStyle,
   BallotType,
   LanguageCode,
@@ -30,7 +30,7 @@ const exampleBallotProps: BaseBallotProps = {
     ...election,
     ballotLayout: {
       ...election.ballotLayout,
-      paperSize: BallotPaperSize.Legal,
+      paperSize: HmpbBallotPaperSize.Legal,
     },
     ballotStyles: [ballotStyle],
   },