## Overview
Update the central scan scanning to:

- Expect a pageHeight of max(bmdThermalPageLength, hmpbPageLength) instead of just hmpbPageLength
- Fill in the image with black background instead of white, this allows hmpb to be cropped as they are today 

Updates the bmd intepretation path to then look everywhere on the page for the qr code. I still break into two "top" and "bottom" search areas as the logic to orient the image relies on that data point. 

Adds a quick benchmark test for bmd ballots. 

Perf numbers from turning on debug logging and looking at the total interpretAndSaveImages times: 

- 22inch ballot election with bmd ballot
  - before: 650ms 
  - after: 880ms
  - change: 230ms
- letter size ballot election with bmd ballot
  - before: 510ms
  - after: 660ms
  - change: 150ms